Store the stop words used by the indexer in the database.
To properly search in the database, a user need to use the same stop word
that the ones used while indexing the content.

By storing the used stop words in the database, a user code can use them
on its side and correctly parse a query.
It is not enough to store the language as we need to read a file/resources
and user code may not have them.

Change-Id: I6cbc9f8d30c39d4fc1e65a356347d8fbfd456494