Add the redirect Articles after "normal" Articles.
This should help for openzim/libzim#31
The problem is that redirect article are added in the article vector
(with all other articles). When cleanning the redirect article, we may
remove invalid redirects, and so copy all articles after it.

By adding the redirect Articles at the end of the list and not at the
beginning, we avoid to many copies because redirect is already at the
end of the vector.