Fix url() issue in CSS files
zimwriterfs parses CSS files to find `url()` instructions and replace its content with an updated in-ZIM URL.
Goal is to replace `url(../img/bg.png)` inside `styles/css/style.css` with `url(../../I/styles/img/bg.png)`.

This happens in `computeNewUrl(string& aid, string& baseUrl, string& targetUrl)` and `computeAbsolutePath(string& path,
string& relativePath)` calls (the first calling the latter).

The problem is that we are calling `computeNewUrl(url, longUrl, path);` where `url` is a local variable containing the image target URL and not the CSS *parent* file URL. Without this information, `computeAbsolutePath(url, path)` is unable to guess where the relative `url()` info starts from.

I am confident this is a typo as it is used properly in `parseAndAdaptHtml()` using `url` but there, `url` is not defined locally and this is brought from `this->url`.