Fix zimdump on windows.
- Do not use symlink on windows.
- Do not use std::ofstream.

It is not possible to create symlink on windows. At least without being
an administrator. So let's remove the functionality
On windows, we always create html redirect and duplicate the content for
non html redirect.

It is not possible to to use a ofstream on windows as we cannot pass a
wstring to the constructor. So we use the plain old fd.