Use double braces to initialize std::array.
std::array will be initialize to an array of 10 elements (0) with
`{10, 0}`.
To initialize it with 2 elements (10 and 0), we use a double braces
`{{10, 0}}`.

To avoid mistake xcode genere a waring and as we are using `-Werror` it
breaks compilation.
(gcc remove this warnig because it was too anoying).
See https://bugs.llvm.org/show_bug.cgi?id=21629