Caching of the link status in ArticleChecker
Access to dirents during the internal url check is the bottleneck for
zimcheck performance, making it IO-bound because of numerous random IO
accesses and significantly limiting gains from multithreading.

Once an internal link is checked by `ArticleChecker` it makes sense to
cache the status of that url. This is different from the dirent caching
in libzim in the following ways:

1. A negative result is also cached. Without `ArticleChecker::linkStatusCache`,
   the same url could be passed to libzim again and the dirent caches
   therein would not be of much help since libzim would still need to look
   at the disk in order to make sure that the dirent didn't magically
   appear since last time it proved to be missing from that archive.
   However, this effect has noticeable impact on performance only if
   a significant portion of internal links are broken.

2. Dirent lookup works via a binary search on the full list of dirents.
   So if we are looking for dirent X, we will arrive at it after
   visiting a number of unrelated dirents that also fight for cache
   slots (though this is alleviated to some extent by a separate
   immutable cache in DirentLookup). `ArticleChecker::linkStatusCache`
   skips the binary search over the full list (of course, only in the
   case if the queried item is in the cache).

3. `ArticleChecker::linkStatusCache` is more memory efficient since
   (taking advantage of the application logic) only the url of the
   dirent is stored in the cache (whereas libzim is a general purpose
   service provider has to cache entire dirents).

!!! SPECIAL NOTE !!!:

This commit introduces copies of libzim's private source files
`concurrent_cache.h` and `lrucache.h` in the zim-tools repository.
We probably need another repository for foundation/general-purpose
utilities that can be shared among all openzim and kiwix projects.