Use search regex instead of match for metadata check.
Fix #345, #352

Investigating #345, it seems that zimcheck-test is getting a segfault
(when run with qemu-aarch64) in `std::regex_match` and the PNG_REGEXP_DATA.
It appears the segfault happens with a huge callstack.
I don't know where the bug is exactly (qemu ? std ?) but the behavior is
the same than #352.

By using `regex_search` we allow the regex to not match the whole content
(but still allow it if the regex ends with `$`)

As we don't need to check all the PNG data, and check only the beginning,
`regex_search` is passing without segfault (early exit ?)